chore: consolidate ordered set schema under one directory
